Move timestamp/duration formatting to utils
We will need them from other places, it's not longer just `execute` who
will be dealign with times.
